fix(ci): add missing INV-9 fixture and extract selectAssemblyConstituents
Two CI failures on PR #169:

- test: Luke-1-08v01.odt was only present locally (test/docs/serverDocs
  is gitignored); force-add it like the other .odt fixtures so the INV-9
  cases in prepareConstituentForAssembly.test.ts can copy it in CI.
- integration: assembleQuarter.integration.test.ts imported
  selectAssemblyConstituents from assemblyController, which transitively
  pulls better-auth's ESM-only entrypoint that the integration Jest
  project cannot load (its contract forbids importing server-app code).
  Move the pure function to src/server/assembly/ alongside the other
  assembly helpers and import it from there in both call sites.
